crewgooglegemini/requirements.txt
crewai_tools
langchain_google_genai
load_dotenv